AzureStack: Merged import tests into the regular tests (VMSS, LoadBalancer, Route, RouteTable) (#1889)
* removed seperate import tests for VMSS

* moved lb import tests to regular ones

* moved reoute & route table import tests to regular ones
