--- azurerm/import_arm_loadbalancer_backend_address_pool_test.go
@@ -1,34 +1,0 @@
-package azurerm
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/hashicorp/terraform/helper/acctest"
-	"github.com/hashicorp/terraform/helper/resource"
-)
-
-func TestAccAzureRMLoadBalancerBackEndAddressPool_importBasic(t *testing.T) {
-	resourceName := "azurerm_lb_backend_address_pool.test"
-
-	ri := acctest.RandInt()
-	addressPoolName := fmt.Sprintf("%d-address-pool", ri)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testCheckAzureRMLoadBalancerDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccAzureRMLoadBalancerBackEndAddressPool_basic(ri, addressPoolName, testLocation()),
-			},
-			{
-				ResourceName:      resourceName,
-				ImportState:       true,
-				ImportStateVerify: true,
-				// location is deprecated and was never actually used
-				ImportStateVerifyIgnore: []string{"location"},
-			},
-		},
-	})
-}
