--- azurerm/import_arm_loadbalancer_nat_pool_test.go
@@ -1,34 +1,0 @@
-package azurerm
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/hashicorp/terraform/helper/acctest"
-	"github.com/hashicorp/terraform/helper/resource"
-)
-
-func TestAccAzureRMLoadBalancerNatPool_importBasic(t *testing.T) {
-	resourceName := "azurerm_lb_nat_pool.test"
-
-	ri := acctest.RandInt()
-	natPoolName := fmt.Sprintf("NatPool-%d", ri)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testCheckAzureRMLoadBalancerDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccAzureRMLoadBalancerNatPool_basic(ri, natPoolName, testLocation()),
-			},
-			{
-				ResourceName:      resourceName,
-				ImportState:       true,
-				ImportStateVerify: true,
-				// location is deprecated and was never actually used
-				ImportStateVerifyIgnore: []string{"location"},
-			},
-		},
-	})
-}
