--- azurerm/import_arm_loadbalancer_nat_rule_test.go
@@ -1,34 +1,0 @@
-package azurerm
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/hashicorp/terraform/helper/acctest"
-	"github.com/hashicorp/terraform/helper/resource"
-)
-
-func TestAccAzureRMLoadBalancerNatRule_importBasic(t *testing.T) {
-	resourceName := "azurerm_lb_nat_rule.test"
-
-	ri := acctest.RandInt()
-	natRuleName := fmt.Sprintf("NatRule-%d", ri)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testCheckAzureRMLoadBalancerDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccAzureRMLoadBalancerNatRule_basic(ri, natRuleName, testLocation()),
-			},
-			{
-				ResourceName:      resourceName,
-				ImportState:       true,
-				ImportStateVerify: true,
-				// location is deprecated and was never actually used
-				ImportStateVerifyIgnore: []string{"location"},
-			},
-		},
-	})
-}
