--- azurerm/import_arm_loadbalancer_probe_test.go
@@ -1,34 +1,0 @@
-package azurerm
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/hashicorp/terraform/helper/acctest"
-	"github.com/hashicorp/terraform/helper/resource"
-)
-
-func TestAccAzureRMLoadBalancerProbe_importBasic(t *testing.T) {
-	resourceName := "azurerm_lb_probe.test"
-
-	ri := acctest.RandInt()
-	probeName := fmt.Sprintf("probe-%d", ri)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testCheckAzureRMLoadBalancerDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccAzureRMLoadBalancerProbe_basic(ri, probeName, testLocation()),
-			},
-			{
-				ResourceName:      resourceName,
-				ImportState:       true,
-				ImportStateVerify: true,
-				// location is deprecated and was never actually used
-				ImportStateVerifyIgnore: []string{"location"},
-			},
-		},
-	})
-}
