+++ azurerm/resource_arm_loadbalancer_probe_test.go
@@ -35,6 +35,13 @@ func TestAccAzureRMLoadBalancerProbe_basic(t *testing.T) {
 						"azurerm_lb_probe.test", "id", probeId),
 				),
 			},
+			{
+				ResourceName:      "azurerm_lb.test",
+				ImportState:       true,
+				ImportStateVerify: true,
+				// location is deprecated and was never actually used
+				ImportStateVerifyIgnore: []string{"location"},
+			},
 		},
 	})
 }
